% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BSM_LAKE_CLIM_MOCK}
\alias{BSM_LAKE_CLIM_MOCK}
\title{Mock lake-climate dataset}
\format{
A data frame with 1000 rows (lakes) and 27 columns containing fake data.
}
\usage{
BSM_LAKE_CLIM_MOCK
}
\description{
Fake climate values for lakes included in the studies.
}
\details{
Each row is a data entry for one lake. The following abbreviations are used
in column names:
\itemize{
\item temp: temperature
\item precip: precipitation
\item GDD: growing degree days
\item GS: growing season
The values at the end of the variable names (10, 15, 20 and 25) correspond
to the length of the temporal window (in years prior to sampling) used to
compute mean values. Finally \code{dist_highway} is the shortest distance to a
highway and \code{angling_intensity_SUM} is the angling pressure.
}
}
\keyword{datasets}
