% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BSM_MASTER_MOCK}
\alias{BSM_MASTER_MOCK}
\title{Mock main dataset}
\format{
A data frame with 130,776 rows and 27 columns containing fake data.
}
\usage{
BSM_MASTER_MOCK
}
\description{
Mock dataset designed to reproduce the analysis without sharing original
data (we are not allowed to distribute the data).
}
\details{
Each row is a data entry for one fish. The table includes data for lakes:
\itemize{
\item Cycle: BSM cycle during which the lake was sample
\item FMZ: Fisheries management zones
\item Wby_LID_Year: Lake identifiers (dummy ones)
\item Wby_LID_Year: Lake identifiers (serial number to ease some data operations)
\item Latitude, Longitude
\item Mean.Depth, Max.Depth: mean and maximum depth of the lake
\item Depth_Stratum
\item Secchi.Depth
\item pH
\item Conductivity
\item Total.Phosphorus
\item Area_ha: lake area
\item pLittoral, SDF: proportion of littoral, Shoreline development factor
\item WC_TDS: Total dissolved solids
}

as well as fish data:
\itemize{
\item Species, SpeciesCode: Species name and code
\item FLEN: Fork length
\item RWT: Round Weight
\item Sex: 1 = male, 2 = female, and 3 (or 4) = unknown
\item Age, AgeConfidence: Species Age and its confidence (Age +/- AgeConfidence)
\item AgingStructure: structure used to determine the age
}
}
\references{
\url{https://www.ontario.ca/page/broad-scale-fish-community-monitoring}
}
\keyword{datasets}
