% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_changing_Efxplot_function.R
\name{Efxplot}
\alias{Efxplot}
\title{Plot function}
\usage{
Efxplot(
  ModelList,
  Sig = TRUE,
  StarLoc = NULL,
  Alpha1 = 1,
  Alpha2 = 1,
  PointOutline = FALSE,
  ModelNames = NULL,
  VarNames = NULL,
  VarOrder = NULL,
  Intercept = TRUE,
  Size = 1,
  tips = 0.2
)
}
\description{
Adapted from ggregplot (Model Effects Plot.R)
}
\references{
\itemize{
\item \url{https://rdrr.io/github/gfalbery/ggregplot/src/R/Model\%20Effects\%20Plot.R}
\item https://rdrr.io/github/gfalbery/ggregplot/src/R/INLA\%20DIC\%20Plot.R
}
}
