% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_create_sup_fig.R
\name{figSup}
\alias{figSup}
\title{Supplementary figure}
\usage{
figSup(bsm_master = readRDS("data/BSM_MASTER.rds"))
}
\arguments{
\item{bsm_master}{main dataset.}
}
\description{
Supplementary figure
}
