% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Ontario_Map.R
\name{mapBSM}
\alias{mapBSM}
\title{Context map}
\usage{
mapBSM(bsm_master = readRDS("data/BSM_MASTER.rds"), useElevation = FALSE)
}
\arguments{
\item{bsm_master}{main dataset.}

\item{useElevation}{a logical. Should a raster of elevation be used as base
map?}
}
\description{
Context map
}
