% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_INLA.R
\name{fitInla}
\alias{fitInla}
\alias{allExfPlot}
\title{fit INLA models}
\usage{
fitInla(nMinIndiv = 15, useMock = FALSE)

allExfPlot(lsMod)
}
\arguments{
\item{nMinIndiv}{minimal number of individuals.}

\item{lsMod}{List of INLA model.}
}
\description{
fit INLA models
}
\section{Functions}{
\itemize{
\item \code{allExfPlot()}: Generates ExfPlot plots.

}}
